+++ hspec-core/test/Test/Hspec/UtilSpec.hs
@@ -28,6 +28,11 @@ spec = do
     it "converts exception to string" $ do
       formatException (E.toException E.DivideByZero) `shouldBe` "ArithException (divide by zero)"
 
+    context "when used with an IOException" $ do
+      it "includes the IOErrorType" $ do
+        Left e <- E.try (readFile "foo")
+        formatException e `shouldBe` "IOException of type NoSuchThing (foo: openFile: does not exist (No such file or directory))"
+
   describe "lineBreaksAt" $ do
     it "inserts line breaks at word boundaries" $ do
       lineBreaksAt 20 "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod"