+++ CHANGES.markdown
@@ -1,6 +1,7 @@
 ## Changes in 2.0.0
   - Split package into `hspec`, `hspec-core` and `hspec-discover`
   - Allow hooks (`before`, `around`, etc.) to pass arguments to spec items
+  - Include `IOErrorType` when printing uncaught `IOException`s (see #204)
 
 ## Changes in 1.12.1
   - Add `--module-name` flag to `hspec-discover` (see #168)